rename package, import gen_ai attributes, fix tests
