+++ python/instrumentation/openinference-instrumentation-pydantic-ai/LICENSE
