+++ python/instrumentation/openinference-instrumentation-pydantic-ai/src/openinference/instrumentation/pydantic_ai/__init__.py
@@ -1,0 +1,9 @@
+from openinference.instrumentation.pydantic_ai.span_exporter import OpenInferenceSpanExporter
+from openinference.instrumentation.pydantic_ai.utils import is_openinference_span
+from openinference.instrumentation.pydantic_ai.version import __version__
+
+__all__ = [
+    "OpenInferenceSpanExporter",
+    "is_openinference_span",
+    "__version__",
+]
