+++ python/instrumentation/openinference-instrumentation-pydantic-ai/tests/openinference/instrumentation/pydanticai/cassettes/TestPydanticAIInstrumentation.test_pydantic_ai_instrumentation.yaml
