+++ python/instrumentation/openinference-instrumentation-pydantic-ai/tests/openinference/instrumentation/pydanticai/test_instrumentor.py
@@ -13,8 +13,8 @@ from pydantic_ai import Agent
 from pydantic_ai.models.openai import OpenAIModel
 from pydantic_ai.providers.openai import OpenAIProvider
 
-from openinference.instrumentation.pydanticai.span_exporter import OpenInferenceSpanExporter
-from openinference.instrumentation.pydanticai.utils import is_openinference_span
+from openinference.instrumentation.pydantic_ai.span_exporter import OpenInferenceSpanExporter
+from openinference.instrumentation.pydantic_ai.utils import is_openinference_span
 from openinference.semconv.trace import (
     MessageAttributes,
     OpenInferenceSpanKindValues,
@@ -59,9 +59,15 @@ class TestPydanticAIInstrumentation:
         )
         assert attributes.get(SpanAttributes.LLM_SYSTEM) == "openai"
         assert attributes.get(SpanAttributes.LLM_MODEL_NAME) == "gpt-4o"
-        assert attributes.get(SpanAttributes.LLM_TOKEN_COUNT_PROMPT) == 58
-        assert attributes.get(SpanAttributes.LLM_TOKEN_COUNT_COMPLETION) == 22
-        assert attributes.get(SpanAttributes.LLM_TOKEN_COUNT_TOTAL) == 80
+
+        prompt_tokens = attributes.get(SpanAttributes.LLM_TOKEN_COUNT_PROMPT)
+        completion_tokens = attributes.get(SpanAttributes.LLM_TOKEN_COUNT_COMPLETION)
+        total_tokens = attributes.get(SpanAttributes.LLM_TOKEN_COUNT_TOTAL)
+        assert isinstance(prompt_tokens, int)
+        assert isinstance(completion_tokens, int)
+        assert isinstance(total_tokens, int)
+        assert total_tokens == prompt_tokens + completion_tokens
+        # Check that tools are present
         assert (
             attributes.get(f"{SpanAttributes.LLM_TOOLS}.0.{SpanAttributes.TOOL_NAME}")
             == "final_result"
