--- python/instrumentation/openinference-instrumentation-pydanticai/MIGRATION.md
@@ -1,1 +1,0 @@
-# Migrations
