--- python/instrumentation/openinference-instrumentation-pydanticai/README.md
@@ -1,3 +1,0 @@
-# OpenInference PydanticAI Instrumentation
-
-Python auto-instrumentation library for PydanticAI.
