--- python/instrumentation/openinference-instrumentation-pydanticai/src/openinference/instrumentation/pydanticai/__init__.py
@@ -1,9 +1,0 @@
-from openinference.instrumentation.pydanticai.span_exporter import OpenInferenceSpanExporter
-from openinference.instrumentation.pydanticai.utils import is_openinference_span
-from openinference.instrumentation.pydanticai.version import __version__
-
-__all__ = [
-    "OpenInferenceSpanExporter",
-    "is_openinference_span",
-    "__version__",
-]
