+++ python/tox.ini
@@ -27,7 +27,7 @@ envlist =
   py3{9,13}-ci-{google_genai,google_genai-latest}
   py39-mypy-langchain_core
   py3{10,13}-ci-{mcp,mcp-latest}
-  py3{9,13}-ci-{pydanticai,pydanticai-latest}
+  py3{9,13}-ci-{pydantic-ai,pydantic-ai-latest}
 
 [testenv]
 package = wheel
@@ -60,7 +60,7 @@ changedir =
   portkey: instrumentation/openinference-instrumentation-portkey/
   mcp: instrumentation/openinference-instrumentation-mcp/
   google_genai: instrumentation/openinference-instrumentation-google-genai/
-  pydanticai: instrumentation/openinference-instrumentation-pydanticai/
+  pydantic-ai: instrumentation/openinference-instrumentation-pydantic-ai/
 commands_pre =
   agno: uv pip install --reinstall {toxinidir}/instrumentation/openinference-instrumentation-agno[test]
   agno-latest: uv pip install -U agno
@@ -132,7 +132,7 @@ commands_pre =
   mcp: python -c 'import openinference.instrumentation.mcp'
   mcp: uv pip install -r test-requirements.txt
   mcp-latest: uv pip install -U mcp
-  pydanticai: uv pip install --reinstall {toxinidir}/instrumentation/openinference-instrumentation-pydanticai[test]
+  pydantic-ai: uv pip install --reinstall {toxinidir}/instrumentation/openinference-instrumentation-pydantic-ai[test]
 
 commands =
   ruff: ruff format .
